Added ability to move cameras from script

+++ Engine/sos.h
@@ -17,7 +17,7 @@ struct transform {
 struct sos_model {
   struct mesh *mesh;
   unsigned short *texture;
-  float rotationAngle;
+  double rotationAngle;
   int visible;
   struct vector3 *position;
   struct vector3 *rotationAxis;
